refactor(control-server): extract shared WebSocket test helpers
recordJsonMessages, messageCollector, and the boot/connect/redeem-invite
boilerplate were each copy-pasted with small variations across
wsValidation.test.ts, multiplexing.test.ts, uplinkAuth.test.ts, and
wsRateLimit.test.ts. Move them into testHelpers.ts as
connectStreamwallUplink, mintUplinkToken, redeemInviteAndConnectClient,
listenTestApp, recordJsonMessages, messageCollector, and VALID_STATE,
and have the four test files import them instead of keeping their own
copies.

No behavior change; the existing test suites (still 90 assertions in
this package, 559 across the repo) are the verification.

Closes #216
